requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
gradio>=3.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
nltk==3.8.1